pyproject.toml
[project]
name = "board"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[tool.poetry]
packages = [{include = "board", from = "src"}]

[tool.poetry.group.staging.dependencies]
django-debug-toolbar = "^5.0.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
